Abstracted the color shift animation to it's own hook.
