+++ pwa/src/pages/CreateTransaction/components/TypeToggleButton/index.tsx
@@ -1,13 +1,8 @@
-import React, { useCallback, useMemo } from 'react'
-import { TextStyle } from 'react-native'
-import Animated, {
-  interpolateColors,
-  EasingNode,
-  useValue,
-} from 'react-native-reanimated'
+import React, { useCallback } from 'react'
 
 import { useNewTransaction } from '../../../../hooks/newTransaction'
 import useToggleButtonAnimation from './useToggleButtonAnimation'
+import useTextColorShiftAnimation from './useTextColorShiftAnimation'
 
 import {
   Container,
@@ -25,39 +20,17 @@ const TypeToggleButton = () => {
     toggleButtonAnimation,
   } = useToggleButtonAnimation(transactionType)
 
-  const colorShift = useValue<number>(transactionType === 'income' ? 0 : 1)
-
-  const colorShiftAnimation = useCallback(() => {
-    Animated.timing(colorShift, {
-      toValue: transactionType === 'income' ? 1 : 0,
-      duration: 150,
-      easing: EasingNode.linear,
-    }).start()
-  }, [colorShift, transactionType])
+  const {
+    incomeTextColorAnimatedStyle,
+    outcomeTextColorAnimatedStyle,
+    colorShiftAnimation,
+  } = useTextColorShiftAnimation(transactionType)
 
   const handleToggleTransactionType = useCallback(() => {
     toggleTransactionType()
     colorShiftAnimation()
     toggleButtonAnimation()
   }, [colorShiftAnimation, toggleButtonAnimation, toggleTransactionType])
-
-  const incomeTextColorAnimatedStyle = useMemo(() => {
-    return ({
-      color: interpolateColors(colorShift, {
-        inputRange: [0, 1],
-        outputColorRange: ['rgb(237,237,237)', 'rgb(59,59,59)'],
-      }),
-    } as unknown) as TextStyle
-  }, [colorShift])
-
-  const outcomeTextColorAnimatedStyle = useMemo(() => {
-    return ({
-      color: interpolateColors(colorShift, {
-        inputRange: [0, 1],
-        outputColorRange: ['rgb(59,59,59)', 'rgb(237,237,237)'],
-      }),
-    } as unknown) as TextStyle
-  }, [colorShift])
 
   return (
     <Container onPress={handleToggleTransactionType}>
